Moves point module to geometry crate
This commit reorganizes the project structure by moving the
point module from `delaunay_core` to a new `geometry` crate.

This change promotes a cleaner separation of concerns, as the
`geometry` crate can now house fundamental geometric primitives
and operations, while `delaunay_core` focuses specifically on
Delaunay triangulation logic.

This refactoring improves code organization and maintainability.
